campaign_site/requirements.txt
Django==2.2.0
djangorestframework==3.9.0
django-phone-field==1.8.1
djangorestframework-simplejwt==4.3.0
djoser==2.0.5
pymysql==0.10.1
kavenegar==1.1.2
zeep==3.4.0
cryptography==3.1.1
pillow==7.0.0
django-cors-headers==3.5.0
django-import-export==2.5.0